Updated Gulp and NpmInstall commands to install node
Updated the Gulp and NpmInstall commands so that they are able to invoke
other commands in the library to ensure that they have all the pieces in
place that they need to successfully execute.

I am not completely thrilled with re-running finalize_options() within
the run() method, but that seems like the most direct way to ensure that
the self.node variable gets correctly initialized once the InstallNode
command has been run.